[crytpo, test-only] Cross-test BLS signing with BLST
Because of our non-standard map-to-curve (see https://github.com/dapperlabs/flow-go/issues/5647), we need a hybrid verification, that uses a standard-compliant map-to-curve (that of Relic) followed by our signing process, and compares the result to that of BLST.
This opens entrypoints to achieve this comparison along the way. This depends on the serde fixes above.